Refactoring Popups. Added Superclass
