+++ main/src/cgeo/geocaching/maps/CachesOverlay.java
@@ -3,7 +3,7 @@ package cgeo.geocaching.maps;
 import cgeo.geocaching.IWaypoint;
 import cgeo.geocaching.Settings;
 import cgeo.geocaching.WaypointPopup;
-import cgeo.geocaching.cgeopopup;
+import cgeo.geocaching.CachePopup;
 import cgeo.geocaching.enumerations.CacheType;
 import cgeo.geocaching.geopoint.Geopoint;
 import cgeo.geocaching.maps.interfaces.CachesOverlayItemImpl;
@@ -228,7 +228,7 @@ public class CachesOverlay extends AbstractItemizedOverlay {
 
             if (StringUtils.isNotBlank(coordinate.getCoordType()) && coordinate.getCoordType().equalsIgnoreCase("cache") && StringUtils.isNotBlank(coordinate.getGeocode())) {
                 CGeoMap.markCacheAsDirty(coordinate.getGeocode());
-                cgeopopup.startActivity(context, coordinate.getGeocode());
+                CachePopup.startActivity(context, coordinate.getGeocode());
             } else if (coordinate.getCoordType() != null && coordinate.getCoordType().equalsIgnoreCase("waypoint") && coordinate.getId() > 0) {
                 CGeoMap.markCacheAsDirty(coordinate.getGeocode());
                 WaypointPopup.startActivity(context, coordinate.getId(), coordinate.getGeocode());
